Move aliases cldr_serde
