+++ provider/cldr/src/cldr_serde/mod.rs
@@ -2,6 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
+pub mod aliases;
 pub mod list_patterns;
 pub mod time_zone_names;
